Extract TypedSymbol class from NodeRepresentation
